Added a separate tasks page for the application
